generate a new version of the "basic" snapshot
Summary:
This checks in a new version of the "basic" snapshot file.

The change I am most interested in picking up is the recent change to make
sure that we always flush column family information in the RocksDB local
store.

I don't think there were too many other consequential changes to our on-disk
data structures since we last regenerated the snapshot in November.  The main
other change is that we no longer make the checkout mount point directories
read-only (D13515854).

Reviewed By: chadaustin

Differential Revision: D14452215

fbshipit-source-id: 87f1cdb61a91ed0f71333f2566ab6b5dac79f628