types: add mock values for Node and Key
Summary: When writing tests, it is often desirable to be able to quickly get a dummy value for a `Node` hash or `Key`. Trying to construct one on the spot can be overly verbose, so let's define some mock values that can be used by tests. This is similar to what Mononoke does (e.g., https://fburl.com/p9u55uye).

Reviewed By: quark-zju

Differential Revision: D14517258

fbshipit-source-id: e3d4cdd60010f44ca681d7a87e6124fe79f8a4c6