infinitepushbackup: use better revset for determining what to back up
Summary:
Change how infinitepush determines what to back up.

Commits to back up are all draft ancestors of non-obsolete commits, *plus* all
draft ancestors of bookmarked commits, which may be obsolete.

Previously, in a stack of the form:
```
  | x B (obsolete, bookmarked)
  | |
  | o A
  |/
  o
```
neither A or B would be backed up, despite normally being visible to the user.

Reviewed By: liubov-dmitrieva

Differential Revision: D14540356

fbshipit-source-id: 0d6ad330c53c818b08f736a9af64704cf0be7cd5