commitcloud: add detection of workspace oscillation
Summary:
Attempt to detect oscillation of commit cloud workspaces by comparing the new
state after the current sync with the state before the previous application of
cloud changes.  If the version number is incremented by 1, the workspace is
brought back to the same state, and less than a minute has passed since the
last time that commit cloud sync ran, abort the current sync step.

This happens after the commits have been backed up, but before the new cloud
workspace is synced to the commit cloud service.  This prevents further
oscillation whilst ensuring the user's commits are still backed up.

Reviewed By: quark-zju

Differential Revision: D14540355

fbshipit-source-id: 20e4b0333f5a7e34b512967a03099625f62ff9d5