hggit: stop git-updatemeta walk when we finish commits with extras
Summary:
Part way through our hg-git repo's history we started adding the git
hash to the extras so it was easier to generate the map files. Let's make
git-updatemeta stop when it passes this boundary.

This is useful for speeding up internal infrastructure that needs to generate
the mapfile from scratch.

Reviewed By: kulshrax

Differential Revision: D14542354

fbshipit-source-id: 7c17fb1b1439f9b4c0c0acf8b5a85790d02a0861