build: set cargo target dir using config file
Summary:
Using cargo for development using mercurial will result in use of random target
directories depending on where cargo is invoked from.

This diff sets the target directory for cargo in the on disk config file so
that all invocations of cargo from our project reuse the previously built
artifacts.

This may make it more problematic to build for multiple targets because cargo
does not recompile when flags change.

Reviewed By: quark-zju

Differential Revision: D14606468

fbshipit-source-id: 56c0b3b192ee94e1ddafe0949cc2267bb995a696