globalrev: introduce command to backfill the globalrev-hg map
Summary:
This is required for the fast commit lookup based on globalrev in
D14687707. This code is heavily inspired from DurhamG's implementation for the
git-hg mapping in the `hggit` extension.

quark-zju has good suggestions for improving this to include the svnrev as well
which can eliminate the requirement for a svnrev-hg mapping in the future. We
can certainly do those in a later iteration.

Reviewed By: quark-zju

Differential Revision: D14674111

fbshipit-source-id: d09d9e65ac0abcafd478c873086898f787398867