mononoke: return HgEntryId in find_path_in_manifest
Summary:
It's cheap so let's return it. It will be used in the next diffs

Context:
There are some commit cloud commits that take ages to compute. For some reason
commit cloud sends too many manifests, which mononoke doesn't filter out and
does a lot of unnecessary checks, which take ages. This stack fixes it.

Reviewed By: farnz

Differential Revision: D14666306

fbshipit-source-id: 75e39f7d88e139f3558825faec5d4e94f6987050