fix the config for bad kernel versions
Summary:
Remove the incorrect regex for `^4.*_fbk13`.  From T34471852 this was
originally only intended to flag 4.11.3-52_fbk13 as bad.  It is incorrectly
flagging `4.16.18-151_fbk13` as a bad release.

This regex isn't needed at all since that release is older than the
`minimum-kernel-version` setting anyway.

It really seems like this config setting should probably just be a single
regex, and shouldn't be split on `,` since you can just use `|` in the regex
instead.  However that seems like a minor issue that should probably be
addressed in a separate diff.

Reviewed By: chadaustin

Differential Revision: D14815156

fbshipit-source-id: b00dd45cb212f6c3e33c02b6216c57d1510a233b