treemanifest: add test demonstrating infinitepush backing up too many trees
Summary:
When infinitepush is backing up merge commits that have a public commit as one
of their parents, and draft commits as the other, treemanifest doesn't look at
the public parents' trees when building the bundle.  This means the bundle may
contain the public parent's trees.  If this is a merge of two distinct
histories, it will contain a full tree for the public parent.

Add a test that shows the generated bundle contains too many trees.

Reviewed By: StanislavGlebik

Differential Revision: D14798683

fbshipit-source-id: 1bc1324cef2c2ade74ff4aa59a391cc0fcccf12f