types: move quickcheck macro imports into tests module
Summary: Previously, this import would trigger dead code warnings because it would be conditionally included with either `cfg(test)` or the `for-tests` feature, whereas the tests module (which was the only user of the import) would only be conditionally included with `cfg(test)`. Fix the warning by moving the import into the tests module itself.

Reviewed By: quark-zju

Differential Revision: D14936954

fbshipit-source-id: ef7a84e8d36645624077283a0fb7798a1746f579