slides: add the segmented changelog slide
Summary:
The segmented changelog idea is somehow interesting. It went through several
iterations. The current form seems good to be checked in.

This time the source is not Markdown, but a Python script to generate tex code.
The generated PDF is small so I checked it in for convenience. It's not too
trivial to get all the build dependencies setup correctly.

The script has some interesting functions like `dag`. It can be reused in
future slides.

Reviewed By: singhsrb

Differential Revision: D14937221

fbshipit-source-id: 539e5b11b6fc67c0f7b393673513580456671fad